Clean up utils and add more comments

+++ ray-operator/controllers/ray/utils/util_test.go
@@ -488,11 +488,11 @@ func TestFindContainerPort(t *testing.T) {
 		},
 	}
 	port := FindContainerPort(&container, "port1", -1)
-	assert.NotEqual(t, port, -1, "expect port1 found")
+	assert.NotEqual(t, port, int32(-1), "expect port1 found")
 	port = FindContainerPort(&container, "port2", -1)
-	assert.NotEqual(t, port, -1, "expect port2 found")
+	assert.NotEqual(t, port, int32(-1), "expect port2 found")
 	port = FindContainerPort(&container, "port3", -1)
-	assert.Equal(t, port, -1, "expect port3 not found")
+	assert.Equal(t, port, int32(-1), "expect port3 not found")
 }
 
 func TestGenerateHeadServiceName(t *testing.T) {
